Reorganize components into src/components, setup searchbar
