+++ src/components/SearchBar/SearchBar.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react'
+import { Text, View } from 'react-native'
+
+export default class SearchBar extends Component {
+  render() {
+    return (
+      <View>
+        <Text> textInComponent </Text>
+      </View>
+    )
+  }
+}
